LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting-api/reporting-api-sends-reports-on-violation.https.sub-expected.txt



PASS Test that image does not load
PASS Event is fired
FAIL Report is observable to ReportingObserver Can't find variable: ReportingObserver
FAIL Violation report status OK. assert_true: No CSP report sent, but expecting one. expected true got false
